Fix #102: Do not overwrite other plugins' `scalacOptions`.
We now maintain a set of `ScalacOptions` "managed" by sbt-tpolecat.
When computing `scalacOptions`, we do not overwrite them, but
instead we get the previous `scalacOptions.value` and we modify it.

We only remove options that are managed by sbt-tpolecat. And we
only add options that are not already there (because they were
already added in an upper delegate scope).

By default, we automatically compute the set of managed options as
all the options that are "ever" added once by sbt-tpolecat in the
delegate chain.